requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
httpx>=0.27.0
python-whois>=0.9.4
python-dotenv>=1.0.0
pydantic>=2.0.0
Pillow>=10.0.0
transformers>=4.38.0
torch>=2.0.0
python-multipart>=0.0.9
aiohttp>=3.9.0
